Split out SlideDeck component
